requirements.txt
pandas>=1.5.0
pyarrow>=10.0.0
numpy>=1.20.0,<2.0.0
opencv-python>=4.5.0,<4.10.0
Pillow>=9.0.0
tqdm>=4.60.0
matplotlib>=3.5.0
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0